# swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Official-Receipt service API endpoints",
    "title": "Official-Receipt API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/health": {
      "get": {
        "summary": "This will check if the server is up",
        "operationId": "GenericRes",
        "responses": {
          "200": {
            "$ref": "#/responses/GenericRes"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      }
    },
    "/notification": {
      "post": {
        "description": "Add notification",
        "tags": [
          "notif"
        ],
        "operationId": "NotificationReq",
        "parameters": [
          {
            "name": "Notification",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NotificationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericRes"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      }
    }
  },
  "definitions": {
    "GenericResponse": {
      "description": "GenericResponse generic response",
      "type": "object",
      "properties": {
        "reason": {
          "type": "string",
          "x-go-name": "Reason"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-notification-service.git/internal/model"
    },
    "JSONError": {
      "description": "JSONError generic response",
      "type": "object",
      "properties": {
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "source": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Source"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-notification-service.git/internal/model/errors"
    },
    "NotificationRequest": {
      "type": "object",
      "properties": {
        "MerchantCode": {
          "type": "string"
        },
        "MerchantKey": {
          "type": "string"
        },
        "Payload": {
          "type": "string"
        },
        "PaymentType": {
          "type": "integer",
          "format": "int64"
        },
        "RefNo": {
          "type": "string"
        },
        "Signature": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-notification-service.git/internal/model"
    }
  },
  "responses": {
    "GenericRes": {
      "description": "GenericRes response model",
      "schema": {
        "$ref": "#/definitions/GenericResponse"
      }
    },
    "JSONErrors": {
      "description": "JSONErrors response model",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/JSONError"
        }
      }
    }
  }
}